+++ intelmq/bots/collectors/http/collector_http.py
@@ -78,6 +78,8 @@ class HTTPCollectorBot(CollectorBot):
         resp = self.session.get(url=http_url)
 
         if resp.status_code // 100 != 2:
+            self.logger.debug('Response headers: %r.', resp.headers)
+            self.logger.debug('Response body: %r.', resp.text)
             raise ValueError('HTTP response status code was %i.' % resp.status_code)
 
         self.logger.info("Report downloaded.")