+++ CHANGELOG.md
@@ -26,7 +26,9 @@ CHANGELOG
 ### Bots
 #### Collectors
 - `intelmq.bots.collectors.eset.collector`: Added (PR#1554 by Mikk Margus Möll).
-- `intelmq.bots.collectors.http.collector_http`: Added PGP signature check functionality (PR#1602 by sinus-x).
+- `intelmq.bots.collectors.http.collector_http`:
+  - Added PGP signature check functionality (PR#1602 by sinus-x).
+  - If status code is not 2xx, the response header and body are logged in debug logging level (#1615).
 
 #### Parsers
 - `intelmq.bots.parsers.eset.parser`: Added (PR#1554 by Mikk Margus Möll).